lab v3.0

+++ include/tbitfield.h
@@ -11,17 +11,18 @@
 #include <iostream>
 #include <cmath>
 #include <cstring>
+#include <algorithm>
 
 using namespace std;
 
-typedef unsigned int TELEM;
+typedef unsigned long long TELEM;
 
 class TBitField
 {
 private:
-  int  BitLen; // длина битового поля - макс. к-во битов
+  int BitLen; // длина битового поля - макс. к-во битов
   TELEM *pMem; // память для представления битового поля
-  int  MemLen; // к-во эл-тов Мем для представления бит.поля
+  int MemLen; // к-во эл-тов Мем для представления бит.поля
 
   // методы реализации
   int   GetMemIndex(const int n) const; // индекс в pМем для бита n       (#О2)
